+++ tests/testthat/test_pipeop_learnercv.R
@@ -96,34 +96,38 @@ test_that("PipeOpLearnerCV - cv ensemble averages fold learners", {
       resampling.predict_method = "cv_ensemble"
     )
   )
+  prob_feature_names = paste0(po$id, ".prob.", task$class_names)
+  response_feature_name = paste0(po$id, ".response")
 
-  trained_task = po$train(list(task))[[1]]
-  expect_setequal(trained_task$feature_names, c(
-    sprintf("%s.response", po$id),
-    paste0(po$id, ".prob.", task$class_names)
-  ))
+  trained_task = po$train(list(task))[[1L]]
+  expect_setequal(trained_task$feature_names, c(response_feature_name, prob_feature_names))
   expect_equal(po$state$predict_method, "cv_ensemble")
   expect_length(po$state$cv_model_states, 2)
 
-  result_task = po$predict(list(task))[[1]]
-  prob_feature_names = paste0(po$id, ".prob.", task$class_names)
+  # Setting seed to make resolving ties deterministic
+  set.seed(1234)
+  result_task = po$predict(list(task))[[1L]]
 
-  pred_probs = as.matrix(result_task$data(rows = task$row_ids, cols = prob_feature_names))
+  pred_probs = as.matrix(result_task$data(cols = prob_feature_names))
   manual_probs = mlr3misc::map(po$state$cv_model_states, function(state) {
     clone = learner$clone(deep = TRUE)
     clone$state = state
     dt = as.data.table(clone$predict(task))
-    data.table::setorder(dt, row_ids)
+    setorder(dt, row_ids)
     as.matrix(dt[, paste0("prob.", task$class_names), with = FALSE])
   })
   manual_prob = Reduce(`+`, manual_probs) / length(manual_probs)
   colnames(manual_prob) = prob_feature_names
+  
   expect_equal(pred_probs, manual_prob)
 
-  result_response = result_task$data(rows = task$row_ids, cols = sprintf("%s.response", po$id))[[1]]
+  result_response = result_task$data(cols = response_feature_name)[[1L]]
+  # Setting seed to make resolving ties deterministic
+  set.seed(1234)
+  manual_response = task$class_names[max.col(manual_prob)]
   expect_equal(
     as.character(result_response),
-    task$class_names[max.col(manual_prob)]
+    manual_response
   )
 })